requirements.txt
imutils==0.5.4
Kivy==2.1.0
kivymd==1.0.2
matplotlib==3.5.1
numpy==1.22.4
Pillow==9.2.0
reverse_geocoder==1.5.1
scikit-image==0.19.2
scipy==1.7.3
sympy==1.10.1
opencv-python